% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createRepo.R
\name{createRepo}
\alias{createRepo}
\title{Create an Empty Local Repository in a Given Directory}
\usage{
createRepo(repoDir, force = TRUE, setAsDefault = TRUE)
}
\arguments{
  \item{repoDir}{A character that specifies the directory
  for the Repository to be made.} \item{force}{If \code{force
  = TRUE} function call forces to create \code{repoDir}
  directory if it did not exist. Default set to \code{force
  = TRUE}.} \item{setAsDefault}{If \code{setAsDefault = TRUE}
  newly created repo is set as the default one.}
}
\description{
\code{createRepo} creates an empty \link{Repository} in a
given directory in which archived artifacts will be stored.
}
\details{
At least one Repository must be initialized before using
other functions from the \pkg{archivist2} package. When
working in groups, you share this directory with other users
(vis dropbox), you can use github repository, or you can use
remote/cloud repository. In local directory all artifacts
are going to be saved in the local Repository folder, with
an SQLite database attached in a file named \code{backpack}.
After calling \code{writeToRepo} function, every artifact
will be archived in a \code{md5hash.rda} file. This file
will be saved in a folder (under \code{repoDir} directory)
named \code{gallery}. For every artifact, \code{md5hash}
is a unique string of length 32 that comes out as a
result of \link[digest]{digest} function, which uses a
cryptographical MD5 hash algorithm. To learn more about
artifacts visit \link[archivist2]{archivist2-package}.
Created \code{backpack} database is a useful and fundamental
tool for remembering artifact's \code{name}, \code{class},
\code{archiving date} etc. (that are remembered as
\link{Tags}), or for keeping artifact's \code{md5hash}.
Besides the \code{backpack} database, \code{gallery} folder
is created in which all artifacts will be archived. After
every \code{writeToRepo} call the database is refreshed, so
an artifact is available immediately in \code{backpack.db}
database for other collaborators.
}
\examples{
\dontrun{
exampleRepoDir <- tempdir()
createRepo( repoDir = exampleRepoDir )

# check the state of an empty Repository
summaryRepo(  repoDir = exampleRepoDir )

# removing an example Repositories
deleteRepo( exampleRepoDir )
rm( exampleRepoDir )
}
}
\author{
Marcin Kosinski, m.p.kosinski_at_gmail.com, Przemyslaw
Biecek, przemyslaw.biecek_at_gmail.com
}
\seealso{
Other archivist2: \code{\link{readFromRepo}};
\code{\link{setRepo}}; \code{\link{writeToRepo}}
}

