% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readFromRepo.R
\name{readFromRepo}
\alias{readFromRepo}
\title{Read an Artifact from a Repository}
\usage{
readFromRepo(md5hash, repo = NULL, value = FALSE)
}
\arguments{
  \item{md5hash}{A character assigned to the artifact as
  a result of a cryptographical hash function with MD5
  algorithm, or it's abbreviation.} \item{repo}{A character
  denoting an existing repository from which an artifact
  will be loaded. If set to \code{NULL} (by default), uses
  the default repository specified in \link{setRepo}.}
  \item{value}{If \code{FALSE} then artifacts are loaded
  into the Global Environment with their original names,
  if \code{TRUE} (default) then artifacts are returned as a
  list of values (if there is more than one artifact) or as
  a single value (if there is only one arfifact that matches
  md5hash).}
}
\description{
\code{readFromRepo} reads an artifact from a local, GitHub
or remote \link{Repository} To learn more about artifacts
visit \link[archivist]{archivist2-package}.
}
\details{
Functions \code{readFromRepo} reads artifacts from the
archivist Repositories stored in a local folder or on
Github or remotely. It takes \code{md5hash} as a hook of
an artifact, which is a result from \link{writeToRepo}
function. For every artifact, \code{md5hash} is a
unique string of length 32 that comes out as a result of
\link[digest]{digest} function, which uses a cryptographical
MD5 hash algorithm. For more information see \link{md5hash}.
Important: instead of giving the whole \code{md5hash}
character, the user can simply give first few characters
of the \code{md5hash}. For example, \code{a09dd} instead
of \code{a09ddjdkf9kj33dcjdnfjgos9jd9jkcv}. All artifacts
with the same corresponing \code{md5hash} abbreviation will
be loaded from \link{Repository}. NOTE: in remote repo only
the first match is returned and warrning is triggered. One
may notice that \code{writeToRepo} remembers the original
artifact name. By default the \code{readFromRepo()} returns
the value, but by setting \code{value = FALSE} you will
force this function to load the artifact directly into the
parent environemnt.
}
\note{
You can specify one \code{md5hash} (or its abbreviation) per
function call.
}
\examples{
\dontrun{
# TODO: create examples or move from archivist
}
}
\author{
Przemyslaw Biecek, przemyslaw.biecek_at_gmail.com, Marcin
Kosinski, m.p.kosinski_at_gmail.com
}
\seealso{
Other archivist2: \code{\link{createRepo}};
\code{\link{setRepo}}; \code{\link{writeToRepo}}
}

