% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setRepo.R
\name{setRepo}
\alias{setRepo}
\title{Set Default Repository}
\usage{
setRepo(localDir = NULL, remoteRepo = NULL, githubUser = NULL,
  githubRepo = NULL, githubBranch = "master", githubDir = "")
}
\arguments{
  \item{localDir}{A path to existing directory with
  Reposiotry.} \item{remoteRepo}{A http:// or https://
  addres of script that accepts POST and GET requests.}
  \item{githubUser}{For github repository it's a github user
  name.} \item{githubRepo}{For github repository it's a github
  repository name.} \item{githubBranch}{For github repository
  it's a github branch (by default master).} \item{githubDir}
  {For github repository it's a github direcotry (by default
  global).}
}
\description{
\code{setRepo} sets default path to the \link{Repository}.
It might be a local repo (directory) GitHub repo (read
only) or remote one. For all \code{archivist2} functions if
\code{repo = NULL} then the default repo is used.
}
\details{
See examples.
}
\examples{
\dontrun{
# examples
# TODO: create examples or move from archivist
}
}
\author{
Przemyslaw Biecek, przemyslaw.biecek_at_gmail.com,
}
\seealso{
\href{https://github.com/pbiecek/archivist2/wiki}{https://
github.com/pbiecek/archivist2/wiki} Other archivist2:
\code{\link{createRepo}}; \code{\link{readFromRepo}};
\code{\link{writeToRepo}}
}

